unordered set filter

+++ src/match_read.cpp
@@ -54,7 +54,7 @@ int get_exact_locs_hash(char* seq, int32_t qpos, uint32_t len, GIMatchedKmer* mk
 	if (cv < 0)
 		return 0;
 
-	if(filtered_kmers.count((hv << (checkSumLength*3)) | cv) > 0)
+	if(filtered_kmers.count(((uint64_t)hv << (checkSumLength*3)) | cv) > 0)
 		return 0;
 
 	GeneralIndex *it = getCandidates(hv);
